requirement.txt
cudatoolkit==11.6.0
h5py==3.7.0
matplotlib-base==3.5.2
matplotlib-inline==0.1.6
numpy-base==1.23.3
numpy==1.23.3
nni==2.10
packaging==21.3
pandas==1.4.4
pickleshare==0.7.5
pillow==9.2.0
pip==22.2.2
plotly==5.9.0
python==3.10.6
python_abi==3.10
pytorch-cluster==1.6.0
pytorch-mutex==1.0
pytorch-scatter==2.0.9
pytorch-sparse==0.6.15
pytorch==1.12.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
torchmetrics==0.10.2

